Changed db from sqlite to postgresql

+++ db/migrate/20140805114738_add_provider_and_uid_to_users.rb
@@ -1,6 +1,6 @@
 class AddProviderAndUidToUsers < ActiveRecord::Migration
   def change
   	add_column :users, :provider, :string
-  	add_column :users, :uid, :integer
+  	add_column :users, :uid, :decimal
   end
 end
